e2e: speed up
While working on adding a new e2e to exercise runtimeConfig options, I
realized that our E2Es are slow partially because we have extended
intervals for `wait` events - at 30s. This is unfortunate as _the first
check_ isn't even called for that 30s period, and meant that the no-op
"wait for object to exist" also took 30s.

This speeds up e2es by removing extra `wait`s, and reducing intervals to
allow tests to make progress faster when not blocked on a slow
network/very slow cpus, without making things _worse_ in those cases.

We also move creating the executor into the test setup, as it's required
for all tests, and we don't currently do per-test namespaces.
